django_perm/permissions_demo/core/serializers.py
from rest_framework import serializers
from .models import Document

class DocumentSerializer(serializers.ModelSerializer):
    class Meta:
        model = Document
        fields = ['id', 'title', 'content', 'owner', 'is_private']
        read_only_fields = ['owner']