django_perm/pyproject.toml
[tool.poetry]
name = "django-perm"
version = "0.1.0"
description = ""
authors = ["Nicolas M. <mine.nicolas@gmail.com>"]
readme = "README.md"
packages = [{include = "django_perm"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.1.5"
djangorestframework = "^3.15.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
